+++ src/main/java/com/example/onlineticketingsystem/controller/RouteController.java
@@ -3,6 +3,9 @@ package com.example.onlineticketingsystem.controller;
 import com.example.onlineticketingsystem.DTO.RouteDTO;
 import com.example.onlineticketingsystem.service.RouteService;
 import org.springframework.beans.factory.annotation.Autowired;
+import org.springframework.http.HttpStatus;
+import org.springframework.http.ResponseEntity;
+import org.springframework.security.core.Authentication;
 import org.springframework.web.bind.annotation.*;
 
 import java.util.List;
@@ -15,23 +18,39 @@ public class RouteController {
     private RouteService routeService;
 
     @PostMapping("/saveRoute")
-    public RouteDTO saveRoute(@RequestBody RouteDTO routeDTO){
-        return routeService.saveRoute(routeDTO);
+    public ResponseEntity<?> saveRoute(@RequestBody RouteDTO routeDTO, Authentication authentication){
+        if (authentication.getAuthorities().stream().noneMatch(grantedAuthority -> grantedAuthority.getAuthority().equals("admin"))){
+            return ResponseEntity.status(HttpStatus.UNAUTHORIZED).body("Access denied. You need to have admin role.");
+        }
+        RouteDTO route = routeService.saveRoute(routeDTO);
+        return ResponseEntity.ok(route);
     }
 
     @GetMapping("/getRoute")
-    public List<RouteDTO> getRoute(){
-        return routeService.getAllRoutes();
+    public ResponseEntity<?> getRoute(Authentication authentication){
+        if (authentication.getAuthorities().stream().noneMatch(grantedAuthority -> grantedAuthority.getAuthority().equals("admin"))){
+            return ResponseEntity.status(HttpStatus.UNAUTHORIZED).body("Access denied. You need to have admin role.");
+        }
+        List<RouteDTO> route = routeService.getAllRoutes();
+        return ResponseEntity.ok(route);
     }
 
     @PutMapping("/updateRoute")
-    public RouteDTO updateRoute(@RequestBody RouteDTO routeDTO){
-        return routeService.updateRoute(routeDTO);
+    public ResponseEntity<?> updateRoute(@RequestBody RouteDTO routeDTO, Authentication authentication){
+        if (authentication.getAuthorities().stream().noneMatch(grantedAuthority -> grantedAuthority.getAuthority().equals("admin"))){
+            return ResponseEntity.status(HttpStatus.UNAUTHORIZED).body("Access denied. You need to have admin role.");
+        }
+        RouteDTO route = routeService.updateRoute(routeDTO);
+        return ResponseEntity.ok(route);
     }
 
 
     @DeleteMapping("/deleteRoute")
-    public boolean deleteRoute(@RequestBody RouteDTO routeDTO){
-        return routeService.deleteRoute(routeDTO);
+    public ResponseEntity<?> deleteRoute(@RequestBody RouteDTO routeDTO, Authentication authentication){
+        if (authentication.getAuthorities().stream().noneMatch(grantedAuthority -> grantedAuthority.getAuthority().equals("admin"))){
+            return ResponseEntity.status(HttpStatus.UNAUTHORIZED).body("Access denied. You need to have admin role.");
+        }
+        Boolean route = routeService.deleteRoute(routeDTO);
+        return ResponseEntity.ok(route);
     }
 }
